+++ external/vulkancts/modules/vulkan/CMakeLists.txt
@@ -33,7 +33,7 @@ add_subdirectory(ycbcr)
 add_subdirectory(protected_memory)
 add_subdirectory(memory_model)
 add_subdirectory(util)
-add_subdirectory(graphicsfuzz)
+add_subdirectory(metamorphic)
 
 include_directories(
 	api
@@ -69,7 +69,7 @@ include_directories(
 	protected_memory
 	memory_model
 	util
-	graphicsfuzz
+	metamorphic
 	)
 
 set(DEQP_VK_SRCS
@@ -124,7 +124,7 @@ set(DEQP_VK_LIBS
 	deqp-vk-ycbcr
 	deqp-vk-protected-memory
 	deqp-vk-memory-model
-	deqp-vk-graphicsfuzz
+	deqp-vk-metamorphic
 	)
 
 if (DE_COMPILER_IS_MSC AND (DE_PTR_SIZE EQUAL 4))
