--- external/vulkancts/modules/vulkan/graphicsfuzz/graphicsFuzzAddTests.inc
@@ -1,2 +1,0 @@
-// Generated
-addFunctionCaseWithPrograms	(graphicsFuzzTests.get(), "mock",	"", createMockProgs, runMock);
