+++ external/vulkancts/modules/vulkan/metamorphic/graphicsFuzzAddTests.inc
@@ -1,0 +1,2 @@
+// Generated
+addFunctionCaseWithPrograms	(metamorphicTests.get(), "mock",	"", createMockProgs, runMock);
