+++ external/vulkancts/modules/vulkan/vktTestPackage.cpp
@@ -87,7 +87,7 @@
 #include "vktProtectedMemTests.hpp"
 #include "vktDeviceGroupTests.hpp"
 #include "vktMemoryModelTests.hpp"
-#include "vktGraphicsFuzzTests.hpp"
+#include "vktMetamorphicTests.hpp"
 
 #include <vector>
 #include <sstream>
@@ -488,7 +488,7 @@ void TestPackage::init (void)
 	addChild(DeviceGroup::createTests		(m_testCtx));
 	addChild(MemoryModel::createTests		(m_testCtx));
 	addChild(conditional::createTests		(m_testCtx));
-	addChild(graphicsfuzz::createTests		(m_testCtx));
+	addChild(metamorphic::createTests		(m_testCtx));
 }
 
 } // vkt
